move definitions to inside of the app process
